REF: Move interfaces from misc to fsl
